# Backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ProjetTutore API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Primary API Server"
        }
    ],
    "paths": {
...
        "/notifications": {
            "get": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Récupérer les notifications (Personnelles + Broadcast)",
                "operationId": "b1274a2aaae8a4295a7e1925e8d8090f",
                "responses": {
                    "200": {
                        "description": "Liste des notifications"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/notifications/unread-count": {
            "get": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Récupérer uniquement le nombre de notifications NON LUES",
                "operationId": "85697ee523a5dace428185c95efb9f2f",
                "responses": {
                    "200": {
                        "description": "Nombre de notifications non lues"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/notifications/{id}/mark-as-read": {
            "patch": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Marquer une notification spécifique comme lue",
                "operationId": "949c12e654c0d6c8830bf691ed7cc54f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notification marquée comme lue"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/notifications/mark-all-as-read": {
            "post": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Marquer TOUTES les notifications comme lues",
                "operationId": "b593c48d89ddf587f9f2f3e522622dcd",
                "responses": {
                    "200": {
                        "description": "Toutes les notifications ont été marquées comme lues"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/notifications": {
            "post": {
                "tags": [
                    "Notifications (Admin)"
                ],
                "summary": "Créer une notification (Admin)",
                "operationId": "3ea51599e555edd0fd445988dfd53a33",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "message"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Titre de la notif"
                                    },
                                    "message": {
                                        "type": "string",
                                        "example": "Message"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "info"
                                    },
                                    "user_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Notification envoyée avec succès"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/notifications/{id}": {
            "delete": {
                "tags": [
                    "Notifications (Admin)"
                ],
                "summary": "Supprimer une notification spécifique (Admin)",
                "operationId": "a580007590de07ab4c7f2827e5f56589",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notification supprimée définitivement"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}